requirements.txt
huggingface-hub
python-dotenv
spacy
google-generativeai>=0.3.0
streamlit>=1.28.0